Move package `smartbar` from `ime.text` to `ime`
