+++ app/src/main/kotlin/dev/patrickgold/florisboard/ime/smartbar/ClipboardCursorRow.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.ime.text.smartbar
+package dev.patrickgold.florisboard.ime.smartbar
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
