+++ app/src/main/kotlin/dev/patrickgold/florisboard/ime/smartbar/InlineSuggestionView.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.ime.text.smartbar
+package dev.patrickgold.florisboard.ime.smartbar
 
 import android.os.Build
 import android.util.Size
